GenAI/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic[email]==2.3.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
email-validator==2.0.0
pytest==7.4.2
httpx>=0.27.0,<0.28.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-ollama>=0.0.1
chromadb>=0.4.22
sentence-transformers>=2.2.2
ollama>=0.1.6
python-multipart>=0.0.6
tenacity>=8.2.3 